remove dqd_emitter
