+++ ribs/emitters/_emitter_base.py
@@ -11,6 +11,8 @@ class EmitterBase(ABC):
     and a :meth:`tell` method that inserts solutions into the emitter's archive.
     Child classes are only required to override :meth:`ask`.
 
+    DQD emitter should override :meth:`ask_dqd` and :meth:`tell_dqd` methods.
+
     Args:
         archive (ribs.archives.ArchiveBase): An archive to use when creating and
             inserting solutions. For instance, this can be
@@ -94,11 +96,12 @@ class EmitterBase(ABC):
         """
         return self._upper_bounds
 
-    @abstractmethod
     def ask(self):
-        """Generates a ``(batch_size, solution_dim)`` array of solutions."""
+        """Generates a ``(batch_size, solution_dim)`` array of solutions.
+        Returns an empty array if not implemented.
+        """
+        return np.empty((0, self.solution_dim))
 
-    @abstractmethod
     def tell(self,
              solution_batch,
              objective_batch,
@@ -126,5 +129,46 @@ class EmitterBase(ABC):
                 single call to archive's :meth:`add()`. For what these floats
                 represent, refer to :meth:`ribs.archives.add()`.
             metadata_batch (numpy.ndarray): 1D object array containing a
+                metadata object for each solution.
+        """
+
+    def ask_dqd(self):
+        """Generates a ``(batch_size, solution_dim)`` array of solutions for
+        which gradient information must be computed, only used for DQD
+        emitters."""
+        return np.empty((0, self.solution_dim))
+
+    def tell_dqd(self,
+                 solution_batch,
+                 objective_batch,
+                 measures_batch,
+                 jacobian_batch,
+                 status_batch,
+                 value_batch,
+                 metadata_batch=None):
+        """Gives the emitter results from evaluating the gradient of the
+        solutions, only used for DQD emitters.
+
+        Args:
+            solution_batch (numpy.ndarray): ``(batch_size,
+                :attr:`solution_dim`)`` array of solutions generated by this
+                emitter's :meth:`ask()` method.
+            objective_batch (numpy.ndarray): 1-dimensional array containing the
+                objective function value of each solution.
+            measures_batch (numpy.ndarray): ``(batch_size, measure space
+                dimension)`` array with the measure space coordinates of each
+                solution.
+            jacobian_batch (numpy.ndarray): ``(batch_size, 1 + measure_dim,
+                solution_dim)`` array consisting of Jacobian matrices of the
+                solutions obtained from :meth:`ask_dqd`. Each matrix should
+                consist of the objective gradient of the solution followed by
+                the measure gradients.
+            status_batch (numpy.ndarray): 1d array of
+                :class:`ribs.archive.addstatus` returned by a series of calls
+                to archive's :meth:`add()` method.
+            value_batch (numpy.ndarray): 1d array of floats returned by a series
+                of calls to archive's :meth:`add()` method. for what these
+                floats represent, refer to :meth:`ribs.archives.add()`.
+            metadata_batch (numpy.ndarray): 1d object array containing a
                 metadata object for each solution.
         """
